chore: restructure mod `group_by_hash`.
